# Conflicts:
#	resources/resources.go

+++ resources/resources.go
@@ -358,7 +358,7 @@ type RequestContext struct {
 
 type VolumeMountProperties struct {
 	WWN          string
-	LunNumber    float64
+	LunNumber    int
 	DeviceMapper string
 	Devices      []string
 }
